+++ harvey/__init__.py
@@ -5,6 +5,7 @@ from harvey.globals import Global
 from harvey.images import Image
 from harvey.messages import Message
 from harvey.pipelines import Pipeline
-from harvey.stages import Stage
+from harvey.stages import (BuildStage, DeployComposeStage, DeployStage,
+                           TestStage)
 from harvey.utils import Logs, Utils
 from harvey.webhooks import Webhook